workaround for pcap_dispatch not receiving
Even though pcap is in non-blocking mode, pcap_dispatch doesn't
reliably return any packets at all when the packet buffer timeout
is zero.

The PCAP_OPEN_LIVE(3PCAP) and PCAP(3PCAP) man pages were
of little to no help in describing this behaviour and it seems
to be a regression in libpcap.